+++ SandboxiePlus/SandMan/Views/SbieView.h
@@ -79,7 +79,8 @@ protected:
 
 	QMap<QString, QStringList>	m_Groups;
 	QSet<QString>				m_Collapsed;
-	//bool						m_UserConfigChanged;
+	bool						m_UserConfigChanged;
+	bool						m_HoldExpand;
 
 private: